api test

+++ app/Http/Controllers/HT/Overview/OverviewController.php
@@ -11,6 +11,7 @@ use App\User;
 //Guzzle
 use GuzzleHttp\Client;
 use Auth;
+use Carbon\Carbon;
 
 class OverviewController extends Controller
 {
@@ -52,15 +53,51 @@ class OverviewController extends Controller
     {
         //dd($request->all());
 
+        //計算該推播的日期   
+        if($request->noticeTime == '分鐘前'){
+            if($request->startTime == null){
+                $push = Carbon::parse($request->start)->subMinutes($request->notice)->format('Y-m-d H:i:s');
+            }
+            else{
+                $push = Carbon::parse($request->start.' '.$request->startTime)->subMinutes($request->notice)->format('Y-m-d H:i:s');
+            }
+        }
+        elseif($request->noticeTime == '小時前'){
+            if($request->startTime == null){
+                $push = Carbon::parse($request->start)->subHours($request->notice)->format('Y-m-d H:i:s');
+            }
+            else{
+                $push = Carbon::parse($request->start.' '.$request->startTime)->subHours($request->notice)->format('Y-m-d H:i:s');
+            }
+        }
+        elseif($request->noticeTime == '天前'){
+            if($request->startTime == null){
+                $push = Carbon::parse($request->start)->subDays($request->notice)->format('Y-m-d H:i:s');
+            }
+            else{
+                $push = Carbon::parse($request->start.' '.$request->startTime)->subDays($request->notice)->format('Y-m-d H:i:s');
+            }
+        }
+        elseif($request->noticeTime == '週前'){
+            if($request->startTime == null){
+                $push = Carbon::parse($request->start)->subWeeks($request->notice)->format('Y-m-d H:i:s');
+            }
+            else{
+                $push = Carbon::parse($request->start.' '.$request->startTime)->subWeeks($request->notice)->format('Y-m-d H:i:s');
+            }
+        }
+
     	 $activity = new Activity;
          $activity->organization_id = $organization->id;
          $activity->user_id = Auth::user()->name;
+         $activity->owner = Auth::user()->id;
     	 $activity->title = $request->title;
     	 ($request->startTime == null)? $activity->start = $request->start : $activity->start = $request->start.' '.$request->startTime;
     	 ($request->endTime == null)? $activity->end = $request->end : $activity->end = $request->end.' '.$request->endTime;
     	 $activity->position = $request->position;
     	 $activity->notice = $request->notice;
          $activity->noticeTime = $request->noticeTime;
+         $activity->pushDate = $push;
     	 $activity->meeting = substr($request->meeting,0,-1);
     	 $activity->description = $request->description;
     	 $activity->save();
@@ -72,16 +109,96 @@ class OverviewController extends Controller
     {
     	$activity = Activity::where('organization_id',Auth::user()->organization_id)->get();
 
+        $date = array();
+        $idArray = array();
+        $titleArray = array();
+        $startArray = array();
+        $endArray = array();
+        $positionArray = array();
+        $meetingArray = array();
+        $noticeArray = array();
+        $noticeTimeArray = array();
+        $descriptionArray = array();
+        $ownerArray = array();
+
+
+        foreach ($activity as $key => $value) {
+            
+           if(!in_array(explode(' ', $value->start)[0],$date) ){
+                array_push($date,explode(' ', $value->start)[0]);
+           }
+        }
+
+        $client = new \GuzzleHttp\Client();
+        $response = $client->get('https://www.googleapis.com/analytics/v3/data/realtime?ids=188946460&metrics=rt:activeUsers');
+
+        $response = $response->getBody()->getContents();
+
+
+        // foreach ($activity as $key => $value) {
+        //     foreach ($date as $k => $v) {
+        //         if(explode(' ', $value->start)[0] == $v){
+        //             $idArray[$v][] = $value->id;
+        //             $titleArray = $value->
+        //             $startArray = $value->
+        //             $endArray = $value->
+        //             $positionArray = $value->
+        //             $meetingArray = $value->
+        //             $noticeArray = $value->
+        //             $noticeTimeArray = $value->
+        //             $descriptionArray = $value->
+        //             $ownerArray = $value->
+        //         }
+        //     }
+        // }
+
+        dd($response);
+
     	return $activity;
     }
 
     public function updateDel(Organization $organization,Request $request)
     {
+
         if (isset($_POST["submit"])) {
             $sub = $_POST["submit"];
 
             if (isset($sub["update"]))
             {
+                //計算該推播的日期   
+                if($request->noticeTime2 == '分鐘前'){
+                    if(isset($request->check)){
+                        $push = Carbon::parse($request->start2)->subMinutes($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                    else{
+                        $push = Carbon::parse($request->start2.' '.$request->startTime2)->subMinutes($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                }
+                elseif($request->noticeTime2 == '小時前'){
+                    if(isset($request->check)){
+                        $push = Carbon::parse($request->start2)->subHours($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                    else{
+                        $push = Carbon::parse($request->start2.' '.$request->startTime2)->subHours($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                }
+                elseif($request->noticeTime2 == '天前'){
+                    if(isset($request->check)){
+                        $push = Carbon::parse($request->start2)->subDays($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                    else{
+                        $push = Carbon::parse($request->start2.' '.$request->startTime2)->subDays($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                }
+                elseif($request->noticeTime2 == '週前'){
+                    if(isset($request->check)){
+                        $push = Carbon::parse($request->start2)->subWeeks($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                    else{
+                        $push = Carbon::parse($request->start2.' '.$request->startTime2)->subWeeks($request->notice2)->format('Y-m-d H:i:s');
+                    }
+                }
+
                 $activity = Activity::find($request->id2);
                 $activity->organization_id = $organization->id;
                 $activity->user_id = Auth::user()->name;
@@ -91,6 +208,7 @@ class OverviewController extends Controller
                 $activity->position = $request->position2;
                 $activity->notice = $request->notice2;
                 $activity->noticeTime = $request->noticeTime2;
+                $activity->pushDate = $push;
 
                 $meeting = implode(",", $request->meeting2);
                 $activity->meeting = $meeting;
@@ -102,9 +220,16 @@ class OverviewController extends Controller
             } 
             elseif (isset($sub["delete"])) 
             {
-               $activity = Activity::find($request->id2)->delete();
+               if(Activity::find($request->id2)->owner == Auth::user()->id){
 
-               return redirect()->route('ht.Overview.index',compact('organization'))->with('success','刪除成功');
+                    $activity = Activity::find($request->id2)->delete();
+
+                    return redirect()->route('ht.Overview.index',compact('organization'))->with('success','刪除成功');
+               }
+               else{
+
+                    return redirect()->route('ht.Overview.index',compact('organization'))->with('error','無此活動刪除權限');
+               }
             }
         }
     }
@@ -125,10 +250,28 @@ class OverviewController extends Controller
 
         $data = json_decode($response);
 
-        $result = array();
-        $bbb = array();
+        $date = array();
+        $typeArray = array();
+        $custkeyArray = array();
+        $addressArray = array();
+        $mobileArray = array();
+        $reasonArray = array();
+        $ownerArray = array();
+        $idArray = array();
+        $dateArray = array();
+        $statusArray = array();
+
+        $typeNew = array();
+        $custkeyNew = array();
+        $addressNew = array();
+        $mobileNew = array();
+        $reasonNew = array();
+        $ownerNew = array();
+        $idNew = array();
+        $dateNew = array();
+        $statusNew = array();
+
         $test = array();
-        $titleArray = array();
 
         foreach ($data as $key => $value) {
             if($key == 'data'){
@@ -137,17 +280,284 @@ class OverviewController extends Controller
         }
 
         foreach ($array as $k => $v) {
-            // $result[$v->time][$k] = ['title'=>$v->id,'start'=>$v->time,'url'=>'#job','className'=>'fc-event-success','allDay'=>'false','id'=>$v->id,'owner'=>$v->owner,'type'=>$v->work_type,'time'=>$v->time,'custkey'=>$v->CUSTKEY,'address'=>$v->address,'mobile'=>$v->mobile,'reason'=>$v->remarks,'status'=>$v->status];
-            if(in_array($v->time, $bbb)){
-                $result[$v->time][] = ['title'=>$v->id];
-            }
-            else{
-                array_push($bbb,$v->time);
-                $result[$v->time][] = ['title'=>$v->id];
+            if(!in_array($v->time, $date)){
+                array_push($date, $v->time);
             }
         }
 
-    	return $response;
+        foreach ($array as $k => $v) {
+            foreach ($date as $kk => $vv) {
+                if($v->time == $vv){
+                    $typeArray[$vv][] = $v->work_type;
+                    $custkeyArray[$vv][] = $v->CUSTKEY;
+                    $addressArray[$vv][] = $v->address;
+                    $mobileArray[$vv][] = $v->mobile;
+                    $reasonArray[$vv][] = $v->remarks;
+                    $ownerArray[$vv][] = $v->owner;
+                    $idArray[$vv][] = $v->id;
+                    $dateArray[$vv][] = $v->time;
+                    $statusArray[$vv][] = $v->status;
+                }
+            }
+        }
+
+        foreach ($typeArray as $type => $types) {
+
+            $test = array();
+
+            foreach ($types as $aa => $aaa) {
+
+                end($types);
+
+                $last = key($types);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $typeNew[$type]['type'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($custkeyArray as $custkey => $custkeys) {
+
+            $test = array();
+
+            foreach ($custkeys as $aa => $aaa) {
+
+                end($custkeys);
+
+                $last = key($custkeys);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $custkeyNew[$custkey]['custkey'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($addressArray as $address => $addresss) {
+
+            $test = array();
+
+            foreach ($addresss as $aa => $aaa) {
+
+                end($addresss);
+
+                $last = key($addresss);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $addressNew[$address]['address'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($mobileArray as $mobile => $mobiles) {
+
+            $test = array();
+
+            foreach ($mobiles as $aa => $aaa) {
+
+                end($mobiles);
+
+                $last = key($mobiles);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $mobileNew[$mobile]['mobile'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($reasonArray as $reason => $reasons) {
+
+            $test = array();
+
+            foreach ($reasons as $aa => $aaa) {
+
+                end($reasons);
+
+                $last = key($reasons);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $reasonNew[$reason]['reason'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($ownerArray as $owner => $owners) {
+
+            $test = array();
+
+            foreach ($owners as $aa => $aaa) {
+
+                end($owners);
+
+                $last = key($owners);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $ownerNew[$owner]['owner'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($idArray as $id => $ids) {
+
+            $test = array();
+
+            foreach ($ids as $aa => $aaa) {
+
+                end($ids);
+
+                $last = key($ids);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $idNew[$id]['number'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($dateArray as $date => $dates) {
+
+            $test = array();
+
+            foreach ($dates as $aa => $aaa) {
+
+                $dateNew[$date]['date'] = $date;
+
+            }
+        }
+
+        foreach ($statusArray as $status => $statuss) {
+
+            $test = array();
+
+            foreach ($statuss as $aa => $aaa) {
+
+                end($statuss);
+
+                $last = key($statuss);
+
+                array_push($test,$aaa);
+
+                if($aa == $last){
+
+                    $resultArray = implode("|||",$test);
+
+                    $statusNew[$status]['status'] = $resultArray;
+                }
+
+            }
+        }
+
+        foreach ($typeNew as $key => $value) {
+            foreach ($custkeyNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $value;
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        foreach ($test as $key => $value) {
+            foreach ($addressNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        foreach ($test as $key => $value) {
+            foreach ($mobileNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        foreach ($test as $key => $value) {
+            foreach ($reasonNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        foreach ($test as $key => $value) {
+            foreach ($ownerNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        foreach ($test as $key => $value) {
+            foreach ($idNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        foreach ($test as $key => $value) {
+            foreach ($dateNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        foreach ($test as $key => $value) {
+            foreach ($statusNew as $k => $v) {
+                if($key == $k){
+                    $test[$key][] = $v;
+                }
+            }
+        }
+
+        $test = array_splice($test, '0'); 
+        $test = array_splice($test, '1'); 
+        $test = array_splice($test, '2'); 
+
+        return $test;
     }
 
     public function getCompany(Organization $organization)
